add: カスタム投稿、カスタムフィールドの実装

+++ src/System/Models/Model.php
@@ -31,20 +31,55 @@ abstract class Model
             $default_value = BindableField::getDefaultValue(get_class($obj), $prop_name);
             if (!is_null($default_value)) {
                 $prop->setValue($obj, $default_value);
+            } elseif ($prop->getType()->allowsNull()) {
+                $prop->setValue($obj, $default_value);
             }
         }
     }
 
+    private static function reverseProperty(Model $obj, ReflectionProperty $prop): string
+    {
+        $prop_name = $prop->getName();
+        if (self::_hasKey($obj, $prop_name)) {
+            $value = self::_getValue($obj, $prop_name);
+            return self::_untypedValue($prop->getType()->getName(), $value);
+        }
+
+        return '';
+    }
+
     private static function _typedValue(string $type, mixed $value): mixed
     {
-        return match ($type) {
-            'bool' => boolval($value),
+        if (is_array($value) && count($value) === 1) {
+            $value = $value[0];
+        }
+        $typed_value = match ($type) {
+            'string' => strval($value),
             'int' => intval($value, 10),
             'float' => floatval($value),
-            'array' => (array) $value,
+            'bool' => boolval($value),
             'DateTime' => DateTimeHelper::datetimeval($value),
             default => $value,
         };
+
+        return $typed_value;
+    }
+
+    private static function _untypedValue(string $type, mixed $value): mixed
+    {
+        if (is_array($value) && count($value) === 1) {
+            $value = $value[0];
+        }
+        $untyped_value = match ($type) {
+            'string' => strval($value),
+            'int' => strval($value),
+            'float' => strval($value),
+            'bool' => strval($value),
+            'DateTime' => DateTimeHelper::strval($value),
+            default => $value,
+        };
+
+        return $untyped_value;
     }
 
     private static function _hasKey(object|array $data, $key): bool
@@ -85,18 +120,33 @@ abstract class Model
     public function register(): int
     {
         $this->ID = wp_insert_post($this);
-
+        $fields = CustomField::getCustomFields(get_class($this));
+        foreach ($fields as $field) {
+            $untypedValue = self::reverseProperty($this, $field);
+            $property = $field->getName();
+            update_post_meta($this->ID, $property, $untypedValue);
+        }
         return $this->ID;
     }
 
     public function update(): void
     {
         wp_update_post($this);
+        $fields = CustomField::getCustomFields(get_class($this));
+        foreach ($fields as $field) {
+            $untypedValue = self::reverseProperty($this, $field);
+            $property = $field->getName();
+            update_post_meta($this->ID, $property, $untypedValue);
+        }
     }
 
     public function delete(bool $force_delete = false): bool
     {
-        $result = wp_delete_post($this->ID, force_delete: $force_delete);
+        if ($force_delete) {
+            $result = wp_delete_post($this->ID, force_delete: $force_delete);
+        } else {
+            $result = wp_trash_post($this->ID);
+        }
         if (!$result) {
             return false;
         }
